refactored GifImage to common module removed dependency to coil from all others
