+++ TODO.md
@@ -9,6 +9,7 @@
 * When user unselect ALL exercise types, allow the session to still run, without showing any
   exercise, as a timer only. Show a message instead of the missing gifs.
 * run ktlintCheck finds issues in a file after auto-format. Fix contradiction between the two
+* some deprecated objects used in the TV version now that it supports part of the compose foundation
 
 ## Publication
