+++ android/mobile/ui/common/src/main/java/fr/shiningcat/simplehiit/android/mobile/ui/common/theme/MobileDarkColorScheme.kt
@@ -1,13 +1,13 @@
 package fr.shiningcat.simplehiit.android.mobile.ui.common.theme
 
 import androidx.compose.material3.darkColorScheme
-import fr.shiningcat.simplehiit.android.common.theme.Amber800
-import fr.shiningcat.simplehiit.android.common.theme.Black
-import fr.shiningcat.simplehiit.android.common.theme.Grey50
-import fr.shiningcat.simplehiit.android.common.theme.Grey800
-import fr.shiningcat.simplehiit.android.common.theme.Grey900
-import fr.shiningcat.simplehiit.android.common.theme.Red_200
-import fr.shiningcat.simplehiit.android.common.theme.Teal600
+import fr.shiningcat.simplehiit.android.common.ui.theme.Amber800
+import fr.shiningcat.simplehiit.android.common.ui.theme.Black
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey50
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey800
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey900
+import fr.shiningcat.simplehiit.android.common.ui.theme.Red_200
+import fr.shiningcat.simplehiit.android.common.ui.theme.Teal600
 
 val mobileDarkColorScheme =
     darkColorScheme(
