+++ android/mobile/ui/common/src/main/java/fr/shiningcat/simplehiit/android/mobile/ui/common/theme/MobileLightColorScheme.kt
@@ -1,14 +1,14 @@
 package fr.shiningcat.simplehiit.android.mobile.ui.common.theme
 
 import androidx.compose.material3.lightColorScheme
-import fr.shiningcat.simplehiit.android.common.theme.Amber500
-import fr.shiningcat.simplehiit.android.common.theme.Black
-import fr.shiningcat.simplehiit.android.common.theme.Grey100
-import fr.shiningcat.simplehiit.android.common.theme.Grey50
-import fr.shiningcat.simplehiit.android.common.theme.Grey900
-import fr.shiningcat.simplehiit.android.common.theme.Red_600
-import fr.shiningcat.simplehiit.android.common.theme.Teal300
-import fr.shiningcat.simplehiit.android.common.theme.White
+import fr.shiningcat.simplehiit.android.common.ui.theme.Amber500
+import fr.shiningcat.simplehiit.android.common.ui.theme.Black
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey100
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey50
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey900
+import fr.shiningcat.simplehiit.android.common.ui.theme.Red_600
+import fr.shiningcat.simplehiit.android.common.ui.theme.Teal300
+import fr.shiningcat.simplehiit.android.common.ui.theme.White
 
 val mobileLightColorScheme =
     lightColorScheme(
