+++ android/mobile/ui/session/src/main/java/fr/shiningcat/simplehiit/android/mobile/ui/session/components/ExerciseDisplayComponent.kt
@@ -13,6 +13,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import fr.shiningcat.simplehiit.android.common.ui.components.GifImage
 import fr.shiningcat.simplehiit.android.mobile.ui.common.theme.SimpleHiitMobileTheme
 import fr.shiningcat.simplehiit.android.mobile.ui.session.CountDown
 import fr.shiningcat.simplehiit.android.mobile.ui.session.RunningSessionStepType
@@ -30,7 +31,6 @@ fun ExerciseDisplayComponent(
     periodType: RunningSessionStepType,
     exerciseSide: ExerciseSide,
     countDown: CountDown? = null,
-    @Suppress("UNUSED_PARAMETER")
     hiitLogger: HiitLogger? = null,
 ) {
     Box(
