+++ android/mobile/ui/settings/build.gradle.kts
@@ -25,8 +25,6 @@ dependencies {
     implementation(libs.androidx.compose.preview)
     implementation(libs.androidx.navigation.compose)
     implementation(libs.androidx.lifecycle)
-    implementation(libs.coil.compose)
-    implementation(libs.coil.gif)
     //
     debugImplementation(libs.androidx.compose.preview.debug)
     debugImplementation(libs.androidx.compose.ui.test.debug)
