+++ android/tv/ui/common/src/main/java/fr/shiningcat/simplehiit/android/tv/ui/common/theme/TvDarkColorScheme.kt
@@ -1,13 +1,13 @@
 package fr.shiningcat.simplehiit.android.tv.ui.common.theme
 
 import androidx.tv.material3.darkColorScheme
-import fr.shiningcat.simplehiit.android.common.theme.Amber800
-import fr.shiningcat.simplehiit.android.common.theme.Black
-import fr.shiningcat.simplehiit.android.common.theme.Grey50
-import fr.shiningcat.simplehiit.android.common.theme.Grey800
-import fr.shiningcat.simplehiit.android.common.theme.Grey900
-import fr.shiningcat.simplehiit.android.common.theme.Red_200
-import fr.shiningcat.simplehiit.android.common.theme.Teal600
+import fr.shiningcat.simplehiit.android.common.ui.theme.Amber800
+import fr.shiningcat.simplehiit.android.common.ui.theme.Black
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey50
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey800
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey900
+import fr.shiningcat.simplehiit.android.common.ui.theme.Red_200
+import fr.shiningcat.simplehiit.android.common.ui.theme.Teal600
 
 val tvDarkColorScheme =
     darkColorScheme(
