+++ android/tv/ui/common/src/main/java/fr/shiningcat/simplehiit/android/tv/ui/common/theme/TvLightColorScheme.kt
@@ -2,14 +2,14 @@ package fr.shiningcat.simplehiit.android.tv.ui.common.theme
 
 import androidx.tv.material3.ColorScheme
 import androidx.tv.material3.lightColorScheme
-import fr.shiningcat.simplehiit.android.common.theme.Amber500
-import fr.shiningcat.simplehiit.android.common.theme.Black
-import fr.shiningcat.simplehiit.android.common.theme.Grey100
-import fr.shiningcat.simplehiit.android.common.theme.Grey50
-import fr.shiningcat.simplehiit.android.common.theme.Grey900
-import fr.shiningcat.simplehiit.android.common.theme.Red_600
-import fr.shiningcat.simplehiit.android.common.theme.Teal300
-import fr.shiningcat.simplehiit.android.common.theme.White
+import fr.shiningcat.simplehiit.android.common.ui.theme.Amber500
+import fr.shiningcat.simplehiit.android.common.ui.theme.Black
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey100
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey50
+import fr.shiningcat.simplehiit.android.common.ui.theme.Grey900
+import fr.shiningcat.simplehiit.android.common.ui.theme.Red_600
+import fr.shiningcat.simplehiit.android.common.ui.theme.Teal300
+import fr.shiningcat.simplehiit.android.common.ui.theme.White
 
 val tvLightColorScheme: ColorScheme
     get() =
