+++ android/tv/ui/statistics/build.gradle.kts
@@ -26,8 +26,6 @@ dependencies {
     implementation(libs.androidx.compose.preview)
     implementation(libs.androidx.navigation.compose)
     implementation(libs.androidx.lifecycle)
-    implementation(libs.coil.compose)
-    implementation(libs.coil.gif)
     //
     debugImplementation(libs.androidx.compose.preview.debug)
     debugImplementation(libs.androidx.compose.ui.test.debug)
